refactor: organize multiplatform backend layout
