+++ 02_quant_dequant/HGinkgo/apps/benchmark_cli.sh
