+++ 02_quant_dequant/HGinkgo/apps/quant_cli.cpp
@@ -1,4 +1,4 @@
-#include "io.hpp"
+#include "low_precision/io.hpp"
 
 #include <cstring>
 #include <cstdlib>
