+++ 02_quant_dequant/HGinkgo/backends/nvidia/cuda_quant.cu
@@ -1,6 +1,6 @@
-#include "cuda_quant.hpp"
+#include "low_precision/cuda_quant.hpp"
 
-#include "quant.hpp"
+#include "low_precision/quant.hpp"
 
 #include <cuda_fp16.h>
 #include <cuda_runtime.h>
