+++ 02_quant_dequant/HGinkgo/common/io.cpp
@@ -1,4 +1,4 @@
-#include "io.hpp"
+#include "low_precision/io.hpp"
 
 #include <array>
 #include <cmath>
