+++ 02_quant_dequant/HGinkgo/tests/test_io.cpp
@@ -1,4 +1,4 @@
-#include "io.hpp"
+#include "low_precision/io.hpp"
 
 #include <cassert>
 #include <cstdio>
