+++ 02_quant_dequant/HGinkgo/tests/test_precision_types.cpp
@@ -1,4 +1,4 @@
-#include "quant.hpp"
+#include "low_precision/quant.hpp"
 
 #include <cassert>
 #include <cmath>
